@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.0</TargetFramework>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
 
     <IsPackable>false</IsPackable>
     <RootNamespace>GameServerLibTests</RootNamespace>
@@ -10,14 +10,16 @@
 
   <ItemGroup>
     <PackageReference Include="nunit" Version="3.12.0" />
-    <PackageReference Include="NUnit3TestAdapter" Version="3.13.0" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.1.1" />
+    <PackageReference Include="NUnit3TestAdapter" Version="3.15.1" />
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.4.0" />
   </ItemGroup>
 
   <ItemGroup>
     <Folder Include="Tests\" />
     <Folder Include="Tests\IO\" />
     <Folder Include="Tests\Protobuf\" />
+    <Folder Include="Tests\Core\" />
+    <Folder Include="Tests\Core\Model\" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\GameServerLib\GameServerLib.csproj" />
