@@ -1,48 +1,33 @@
-﻿/*
-Copyright 2015 Pim de Witte All Rights Reserved.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-using UnityEngine;
-using System.Collections;
+﻿using UnityEngine;
 using System.Collections.Generic;
 using System;
 
-/// Author: Pim de Witte (pimdewitte.com) and contributors, https://github.com/PimDeWitte/UnityMainThreadDispatcher
-/// <summary>
-/// A thread-safe class which holds a queue with actions to execute on the next Update() method. It can be used to make calls to the main thread for
-/// things such as UI Manipulation in Unity. It was developed for use in combination with the Firebase Unity plugin, which uses separate threads for event handling
-/// </summary>
 public class UnityMainThreadDispatcher : MonoBehaviour {
-    public static UnityMainThreadDispatcher instance { get; private set; }
+    private static UnityMainThreadDispatcher _instance;
+
+    public static UnityMainThreadDispatcher instance {
+        get {
+            if (_instance == null) {
+                _instance = new GameObject("MainThreadDispatcher").AddComponent<UnityMainThreadDispatcher>();
+            }
+            return _instance;
+        }
+    }
 
     private readonly Queue<Action> executionQueue = new Queue<Action>();
 
     protected virtual void Awake() {
-        if (instance == null) {
-            instance = this;
-            DontDestroyOnLoad(this.gameObject);
-        } else {
-            Destroy(this.gameObject);
+        if (_instance == null) {
+            _instance = this;
         }
+        DontDestroyOnLoad(this.gameObject);
     }
 
     protected virtual void OnDestroy() {
-        instance = null;
+        _instance = null;
     }
 
-    protected virtual void Update() {
+    public virtual void Update() {
         lock (executionQueue) {
             while (executionQueue.Count > 0) {
                 executionQueue.Dequeue().Invoke();
