@@ -4,7 +4,7 @@ namespace GameNetworking.Commons {
     public abstract class BaseWorker<GameType> where GameType: class, IGameInstance {
         private readonly WeakReference weakServer;
 
-        protected GameType Instance => this.weakServer?.Target as GameType;
+        protected GameType instance => this.weakServer?.Target as GameType;
 
         protected BaseWorker(GameType instance) {
             this.weakServer = new WeakReference(instance);
