@@ -5,11 +5,11 @@ using Messages.Models;
 namespace GameNetworking.Models {
     
     public sealed class NetworkClient {
-        internal NetClient Client { get; private set; }
+        internal INetClient Client { get; private set; }
         internal IStreamReader Reader { get; private set; }
         internal IStreamWriter Writer { get; private set; }
 
-        internal NetworkClient(NetClient client, IStreamReader reader, IStreamWriter writer) {
+        internal NetworkClient(INetClient client, IStreamReader reader, IStreamWriter writer) {
             this.Client = client;
             this.Reader = reader;
             this.Writer = writer;
@@ -23,7 +23,7 @@ namespace GameNetworking.Models {
             if (obj is NetworkClient) {
                 return this.Client == ((NetworkClient)obj).Client;
             }
-            if (obj is NetClient) {
+            if (obj is INetClient) {
                 return this.Client == obj;
             }
             return object.Equals(this, obj);
