@@ -8,46 +8,46 @@ namespace GameNetworking.Models {
 
             private WeakReference weakGameObject;
 
-            internal NetworkClient Client {
+            internal NetworkClient client {
                 get; private set;
             }
 
-            public int PlayerId {
+            public int playerId {
                 get; private set;
             }
 
-            public int SpawnId {
+            public int spawnId {
                 get; set;
             }
 
-            public GameObject GameObject {
+            public GameObject gameObject {
                 get { return this.weakGameObject?.Target as GameObject; }
                 internal set { this.weakGameObject = new WeakReference(value); }
             }
 
-            public Transform Transform {
-                get { return this.GameObject?.transform; }
+            public Transform transform {
+                get { return this.gameObject?.transform; }
             }
 
             internal InputState inputState = new InputState();
 
             internal NetworkPlayer(NetworkClient client) {
-                this.PlayerId = random.Next();
-                this.Client = client;
+                this.playerId = random.Next();
+                this.client = client;
             }
 
             internal NetworkPlayer(int playerId) {
-                this.PlayerId = playerId;
+                this.playerId = playerId;
             }
 
             public void Despawn() {
-                GameObject.Destroy(this.GameObject);
-                this.GameObject = null;
+                GameObject.Destroy(this.gameObject);
+                this.gameObject = null;
             }
 
             public override bool Equals(object obj) {
                 if (obj is NetworkPlayer) {
-                    return this.PlayerId == ((NetworkPlayer)obj).PlayerId;
+                    return this.playerId == ((NetworkPlayer)obj).playerId;
                 }
                 return object.Equals(this, obj);
             }
@@ -59,7 +59,7 @@ namespace GameNetworking.Models {
     }
 
     namespace Client {
-        public class NetworkPlayer: Server.NetworkPlayer {
+        public class NetworkPlayer : Server.NetworkPlayer {
             public bool IsLocalPlayer {
                 get; internal set;
             }
