@@ -6,6 +6,7 @@ namespace GameNetworking {
     using Executors.Client;
     using Executors;
     using Commons;
+    using System.Threading;
 
     internal class GameClientMessageRouter : BaseWorker<GameClient> {
         internal GameClientMessageRouter(GameClient client) : base(client) { }
@@ -18,25 +19,13 @@ namespace GameNetworking {
             if (container == null) { return; }
 
             switch ((MessageType)container.Type) {
-            case MessageType.CONNECTED_PLAYER:
-                Execute(new ConnectedPlayerExecutor(this.Instance, container.Parse<ConnectedPlayerMessage>()));
-                break;
-            case MessageType.SPAWN_REQUEST:
-                Execute(new PlayerSpawnExecutor(this.Instance, container.Parse<PlayerSpawnMessage>()));
-                break;
-            case MessageType.SYNC:
-                Execute(new SyncExecutor(this.Instance, container.Parse<SyncMessage>()));
-                break;
-            case MessageType.PING:
-                Execute(new PingRequestExecutor(this.Instance));
-                break;
-            case MessageType.PING_RESULT:
-                Execute(new PingResultRequestExecutor(this.Instance, container.Parse<PingResultRequestMessage>()));
-                break;
-
-            default:
-                this.Instance?.Delegate?.GameClientDidReceiveMessage(container);
-                break;
+            case MessageType.CONNECTED_PLAYER: this.Execute(new ConnectedPlayerExecutor(this.instance, container.Parse<ConnectedPlayerMessage>())); break;
+            case MessageType.SPAWN_REQUEST: this.Execute(new PlayerSpawnExecutor(this.instance, container.Parse<PlayerSpawnMessage>())); break;
+            case MessageType.SYNC: this.Execute(new SyncExecutor(this.instance, container.Parse<SyncMessage>())); break;
+            case MessageType.PING: this.Execute(new PingRequestExecutor(this.instance)); break;
+            case MessageType.PING_RESULT: this.Execute(new PingResultRequestExecutor(this.instance, container.Parse<PingResultRequestMessage>())); break;
+            case MessageType.DISCONNECTED_PLAYER: this.Execute(new DisconnectedPlayerExecutor(this.instance, container.Parse<DisconnectedPlayerMessage>())); break;
+            default: this.instance?.listener?.GameClientDidReceiveMessage(container); break;
             }
         }
     }
