@@ -1,15 +1,15 @@
 using UnityEngine;
 
 namespace GameNetworking {
-    public interface IGameClientInstanceDelegate {
-        bool GameInstanceSyncPlayer(Models.Client.NetworkPlayer player, Vector3 position, Vector3 eulerAngles);
-    }
-
     public interface IGameInstance {
 
     }
 
-    public interface IGameClientInstance: IGameInstance {
-        IGameClientInstanceDelegate InstanceDelegate { get; set; }
+    public interface IGameClientInstanceListener {
+        bool GameInstanceSyncPlayer(Models.Client.NetworkPlayer player, Vector3 position, Vector3 eulerAngles);
+    }
+
+    public interface IGameClientInstance : IGameInstance {
+        IGameClientInstanceListener instanceListener { get; set; }
     }
 }
