@@ -2,39 +2,31 @@
 using UnityEngine;
 
 namespace GameNetworking.Messages {
-    public class Vec3: ICodable {
-        public float x;
-        public float y;
+    public class Vec3: Vec2 {
         public float z;
 
-        void IDecodable.Decode(IDecoder decoder) {
-            this.x = decoder.Float();
-            this.y = decoder.Float();
+        public override void Decode(IDecoder decoder) {
+            base.Decode(decoder);
             this.z = decoder.Float();
         }
 
-        void IEncodable.Encode(IEncoder encoder) {
-            encoder.Encode(this.x);
-            encoder.Encode(this.y);
+        public override void Encode(IEncoder encoder) {
+            base.Encode(encoder);
             encoder.Encode(this.z);
         }
 
-        public void CopyToVector3(ref Vector3 instance) {
-            instance.x = this.x;
-            instance.y = this.y;
-            instance.z = this.z;
+        public static implicit operator Vector3(Vec3 v) {
+            var vec = Vector3.zero;
+            vec.Set(v.x, v.y, v.z);
+            return vec;
+        }
+
+        public static implicit operator Vec3(Vector3 v) {
+            return new Vec3 { x = v.x, y = v.y, z = v.z };
         }
 
         public override string ToString() {
             return string.Format("({0}, {1}, {2})", this.x, this.y, this.z);
-        }
-    }
-
-    public static class Vector3Ext {
-        public static void CopyToVec3(this Vector3 op, ref Vec3 instance) {
-            instance.x = op.x;
-            instance.y = op.y;
-            instance.z = op.z;
         }
     }
 }
