@@ -3,36 +3,44 @@
     using Models.Server;
     using Messages.Server;
     using Commons;
+    using Logging;
 
-    internal class GameServerClientAcceptor: BaseWorker<GameServer> {
+    internal class GameServerClientAcceptor : BaseWorker<GameServer> {
         public GameServerClientAcceptor(GameServer server) : base(server) { }
 
         public void AcceptClient(NetworkClient client) {
             NetworkPlayer player = new NetworkPlayer(client);
-            this.Instance.SendBroadcast(new ConnectedPlayerMessage { playerId = player.PlayerId, isMe = false });
-            this.Instance.AddPlayer(player);
-            this.Instance.Send(new ConnectedPlayerMessage { playerId = player.PlayerId, isMe = true }, client);
+            this.instance.AddPlayer(player);
+            
+            Logger.Log($"(AcceptClient) count {this.instance.AllPlayers().Count}");
 
-            Logging.Logger.Log(this.GetType(), $"(AcceptClient) count {this.Instance.AllPlayers().Count}");
 
-            this.Instance.AllPlayers().ForEach(each => {
-                if (each.Client == client) { return; }
-                var connected = new ConnectedPlayerMessage();
-                connected.playerId = each.PlayerId;
-                connected.isMe = false;
-                this.Instance.Send(connected, client);
+            this.instance.AllPlayers().ForEach(each => {
+                // Sends the connected player message to all players
+                this.instance.Send(new ConnectedPlayerMessage {
+                    playerId = player.playerId,
+                    isMe = (player == each)
+                }, each.client);
+
+                if (each == player) { return; }
+
+                // Sends the existing players to the player that just connected
+                this.instance.Send(new ConnectedPlayerMessage {
+                    playerId = each.playerId,
+                    isMe = false
+                }, player.client);
             });
         }
 
         public void Disconnect(NetworkClient client) {
-            var player = this.Instance.FindPlayer(client);
-            this.Instance.RemovePlayer(player);
+            var player = this.instance.FindPlayer(client);
+            this.instance.RemovePlayer(player);
 
-            Logging.Logger.Log(this.GetType(), $"(Disconnect) count {this.Instance.AllPlayers().Count}");
+            Logger.Log($"(Disconnect) count {this.instance.AllPlayers().Count}");
 
-            if (player != null) { 
-                this.Instance.Delegate?.GameServerPlayerDidDisconnect(player);
-                this.Instance.SendBroadcast(new DisconnectedPlayerMessage() { playerId = player.PlayerId });
+            if (player != null) {
+                this.instance.listener?.GameServerPlayerDidDisconnect(player);
+                this.instance.SendBroadcast(new DisconnectedPlayerMessage { playerId = player.playerId });
             }
         }
     }
