@@ -1,14 +1,15 @@
 ﻿namespace MatchMaking.Models {
     using Coders;
+    using Networking.Models;
 
     public class Client {
-        internal Networking.Models.NetClient client;
+        internal INetClient client;
         internal IMessageDecoder decoder;
         internal IMessageEncoder encoder;
 
-        public bool IsConnected { get { return this.client.IsConnected; } }
+        public bool IsConnected { get { return this.client.isConnected; } }
 
-        internal static MMClient Create<MMClient>(Networking.Models.NetClient client, IMessageDecoder decoder, IMessageEncoder encoder) where MMClient: Client, new() {
+        internal static MMClient Create<MMClient>(INetClient client, IMessageDecoder decoder, IMessageEncoder encoder) where MMClient: Client, new() {
             return new MMClient {
                 client = client,
                 decoder = decoder,
