@@ -4,27 +4,26 @@ using System.Threading;
 
 namespace Networking.IO {
     public sealed class NetworkingWriter : IWriter {
+        private readonly ISocket socket;
+
         private byte[] buffer;
         private bool isSending;
 
-        private readonly Socket socket;
-
-        internal NetworkingWriter(Socket socket) {
+        internal NetworkingWriter(ISocket socket) {
             this.socket = socket;
             this.buffer = new byte[0];
         }
 
         private void Write() {
-            if (this.socket.Connected) {
+            if (this.socket.isConnected) {
                 if (this.isSending) { return; }
 
                 this.isSending = true;
 
-                this.socket.BeginSend(this.buffer, 0, this.buffer.Length, SocketFlags.Partial, (ar) => {
-                    int written = this.socket.EndSend(ar);
+                this.socket.Write(this.buffer, (written) => {
                     if (written > 0) { this.ShrinkBuffer(written); }
                     this.isSending = false;
-                }, this);
+                });
             }
         }
 
@@ -48,7 +47,7 @@ namespace Networking.IO {
 
     namespace Extensions {
         public static partial class SocketExt {
-            public static IWriter Writer(this Socket op) {
+            public static IWriter Writer(this ISocket op) {
                 return new NetworkingWriter(op);
             }
         }
