@@ -1,5 +1,4 @@
-﻿using System;
-using System.Net;
+﻿using System.Net;
 using System.Net.Sockets;
 using System.Collections.Generic;
 
@@ -7,44 +6,47 @@ namespace Networking {
     using Commons;
     using Models;
     using IO.Extensions;
+    using Networking.IO;
+    using Logging;
 
-    public sealed class NetSocket : WeakDelegate<INetworkingDelegate>, INetworking {
-        private readonly Socket socket;
-        private readonly Queue<Socket> acceptedQueue;
+    public sealed class NetSocket : WeakListener<INetworkingListener>, INetworking {
+        private readonly ISocket socket;
+        private readonly Queue<ISocket> acceptedQueue;
+        private bool isAccepting = false;
 
         public int Port { get; private set; }
 
-        public NetSocket() {
-            this.acceptedQueue = new Queue<Socket>();
-
-            this.socket = new Socket(AddressFamily.InterNetwork, SocketType.Stream, ProtocolType.Tcp) {
-                NoDelay = true,
-                Blocking = false,
-                SendTimeout = 2000,
-                ReceiveTimeout = 2000
-            };
+        public NetSocket(ISocket socket) {
+            this.socket = socket;
+            this.acceptedQueue = new Queue<ISocket>();
         }
 
-        public void Start(int port) {
+        public void StartServer(int port) {
             this.Port = port;
-            this.socket.Bind(new IPEndPoint(IPAddress.Any, port));
+            NetEndPoint ep = new NetEndPoint(IPAddress.Any.ToString(), port);
+            this.socket.Bind(ep);
             this.socket.Listen(10);
-
-            this.AcceptNewClient();
         }
 
         private void AcceptNewClient() {
-            this.socket.BeginAccept((ar) => {
-                var accepted = this.socket.EndAccept(ar);
-                accepted.NoDelay = true;
-                accepted.Blocking = false;
-                this.acceptedQueue.Enqueue(accepted);
+            if (this.isAccepting) { return; }
 
-                this.AcceptNewClient();
-            }, this);
+            this.isAccepting = true;
+
+            this.socket.Accept((accepted) => {
+                if (accepted != null) {
+                    accepted.noDelay = true;
+                    accepted.blocking = false;
+                    this.acceptedQueue.Enqueue(accepted);
+                }
+
+                this.isAccepting = false;
+            });
         }
 
-        public NetClient Accept() {
+        public INetClient Accept() {
+            this.AcceptNewClient();
+
             if (this.acceptedQueue.Count > 0) {
                 var accepted = this.acceptedQueue.Dequeue();
                 return new NetClient(accepted, accepted.Reader(), accepted.Writer());
@@ -53,44 +55,41 @@ namespace Networking {
         }
 
         public void Stop() {
-            this.socket.Shutdown(SocketShutdown.Both);
-            this.socket.Dispose();
+            this.socket.Close();
         }
 
         public void Connect(string host, int port) {
-            var result = this.socket.BeginConnect(host, port, (ar) => {
-                if (this.socket.Connected) {
-                    this.socket.EndConnect(ar);
-                    this.Delegate?.NetworkingDidConnect(new NetClient(this.socket, this.socket.Reader(), this.socket.Writer()));
+            NetClient client = new NetClient(this.socket, this.socket.Reader(), this.socket.Writer());
+            NetEndPoint ep = new NetEndPoint(host, port);
+            client.Connect(ep, () => {
+                if (client.isConnected) {
+                    this.listener?.NetworkingDidConnect(client);
                 } else {
                     this.socket.Close();
-                    this.Delegate?.NetworkingConnectDidTimeout();
+                    this.listener?.NetworkingConnectDidTimeout();
                 }
-            }, null);
+            });
 
-            Logging.Logger.Log(this.GetType(), "Trying to connect to " + host + "-" + port);
+            Logger.Log($"Trying to connect to {host}-{port}");
         }
 
-        public void Disconnect(NetClient client) {
-            client.socket.BeginDisconnect(false, (ar) => {
-                client.socket.EndDisconnect(ar);
-                this.Delegate?.NetworkingDidDisconnect(client);
-            }, this);
+        public void Disconnect(INetClient client) {
+            client.Disconnect(() => this.listener?.NetworkingDidDisconnect(client));
         }
 
-        public void Read(NetClient client) {
+        public void Read(INetClient client) {
             client.reader.Read();
         }
 
-        public void Send(NetClient client, byte[] message) {
+        public void Send(INetClient client, byte[] message) {
             client.writer.Write(message);
         }
 
-        public void Flush(NetClient client) {
-            if (client.IsConnected) {
+        public void Flush(INetClient client) {
+            if (client.isConnected) {
                 client.writer.Flush();
             } else {
-                this.Delegate?.NetworkingDidDisconnect(client);
+                this.listener?.NetworkingDidDisconnect(client);
             }
         }
     }
